[customer_testing] Refactor runner to be testable (#76831)
