+++ dev/devicelab/pubspec.yaml
@@ -35,7 +35,6 @@ dependencies:
   web_socket_channel: 1.2.0 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
 
 dev_dependencies:
-  # See packages/flutter_test/pubspec.yaml for why we're pinning this version.
   test: 1.16.5
 
   _fe_analyzer_shared: 14.0.0 # THIS LINE IS AUTOGENERATED - TO UPDATE USE "flutter update-packages --force-upgrade"
